+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/eva/data/EvaCodepointParser.kt
@@ -1,0 +1,8 @@
+package io.github.composegears.valkyrie.ui.screen.webimport.standard.eva.data
+
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.data.RegexCssCodepointParser
+
+class EvaCodepointParser :
+    RegexCssCodepointParser(
+        Regex("""\.eva-([a-z0-9-]+)::before\s*\{\s*content\s*:\s*"\\([A-Fa-f0-9]+)"\s*;?\s*}"""),
+    )